examples/subobc: apply direcotory refactor migration(Drivers -> driver)
