+++ examples/subobc/src/src_user/applications/driver_instances/di_mobc.h
@@ -5,7 +5,7 @@
 #ifndef DI_MOBC_H_
 #define DI_MOBC_H_
 
-#include "../../Drivers/Etc/mobc.h"
+#include "../../driver/etc/mobc.h"
 #include <src_core/system/application_manager/app_info.h>
 #include <src_core/tlm_cmd/common_cmd_packet.h>
