+++ examples/subobc/src/src_user/driver/etc/mobc.h
@@ -6,7 +6,7 @@
 #define MOBC_H_
 
 #include <src_core/hal/uart.h>
-#include <src_core/Drivers/Super/driver_super.h>
+#include <src_core/driver/driver_super.h>
 #include <src_core/system/time_manager/obc_time.h>
 #include <src_core/tlm_cmd/common_tlm_packet.h>
 #include <src_core/tlm_cmd/packet_handler.h>
